Changed int representation of map stat properties to long
git-svn-id: http://hazelcast.googlecode.com/svn/trunk@1205 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07

+++ hazelcast-monitor/src/main/java/com/hazelcast/monitor/client/MapEntry.java
@@ -30,7 +30,7 @@ public class MapEntry implements Map.Entry, IsSerializable, Serializable {
     private Date lastUpdateTime;
     private Date creationTime;
     private long version = 0;
-    private int hits = 0;
+    private long hits = 0;
     private boolean valid = true;
     private String key = null;
     private String value = null;
@@ -97,11 +97,11 @@ public class MapEntry implements Map.Entry, IsSerializable, Serializable {
         this.version = version;
     }
 
-    public int getHits() {
+    public long getHits() {
         return hits;
     }
 
-    public void setHits(int hits) {
+    public void setHits(long hits) {
         this.hits = hits;
     }
